requirements.txt
pyahocorasick
lxml